Changed error cancel to have paypal in the URL. Went back to full page reload
