+++ src/app/(main)/heartland/PaypalForm.tsx
@@ -21,38 +21,12 @@ export default function PaypalForm() {
   const { orderNumber, billingAddress } = useCheckoutContext();
   const { getTotalPrice } = useCartContext();
   const totalMsrpPrice = getTotalPrice().toFixed(2) as unknown as number;
-  const iframeRef = useRef(null);
-
-  // useEffect(() => {
-  //   // Reload the iframe when the component mounts
-  //   if (iframeRef.current) {
-  //     console.log("Tryin to reload the iframe")
-  //     iframeRef.current.src += '';
-  //   }
-  // }, []);
-
-  // useEffect(() => {
-  //   const handlePageShow = () => {
-  //     if (iframeRef.current) {
-  //       // Reload the iframe to ensure it is in a functional state
-  //       iframeRef.current.src = iframeRef.current.src;
-  //     }
-  //   };
-
-  //   window.addEventListener('pageshow', handlePageShow);
-
-  //   return () => {
-  //     window.removeEventListener('pageshow', handlePageShow);
-  //   };
-  // }, []);
 
   useEffect(() => {
     const handlePageShow = (event) => {
       if (event.persisted) {
         // Force a full page reload if the page was loaded from the cache
-        document
-          .getElementById('paypal-iframe')
-          .contentWindow.location.reload();
+        window.location.reload();
       }
     };
 
@@ -114,7 +88,6 @@ export default function PaypalForm() {
         </div>
       ) : (
         <iframe
-          id="paypal-iframe"
           src={`${payflowUrl}/?SECURETOKEN=${secureToken}&SECURETOKENID=${secureTokenId}`}
           width="600"
           height="550"
